pyproject.toml
[tool.poetry]
name = "llm-agent-project"
version = "0.1.0"
description = ""
authors = ["Konstantin.Grotov <konstantin.grotov@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.34"
setuptools = "^75.1.0"
buildtools = "^1.0.6"
newsapi-python = "^0.2.7"
pandas = "^2.2.3"
pytelegrambotapi = "^4.23.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-telegram-bot = "^20.6"
openai = "^1.3.0"
httpx = "^0.25.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pre-commit = "^4.0.1"


[tool.poetry.group.research.dependencies]
jupyter = "^1.1.1"
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
